requirements.txt
rembg==2.0.67
Pillow==10.4.0
flask==3.0.0
werkzeug==3.0.1
onnxruntime==1.22.1
